Fixing a bug that prevented us from marking versions as processed
We were reusing a command that wrote a new version, but that command
was NOT supposed to assume that the version had been processed
already. That would have prevented download of the plugin at the
plugin download stage.

So I created a new method to explicitly write a new version and mark
it as processed.